feat: improvement `qadiriyah`
[Documentation]

adding Google Test
Google Test (also known as gtest) is a unit testing library for the C++ programming language, released under the BSD 3-clause license[2] and based on the xUnit architecture.[3] Google Test is able to be compiled for a variety of POSIX and Microsoft Windows platforms, allowing unit-testing of C sources as well as C++ with minimal source modification.

adding clang-format
clang-format is a tool to format C/C++/... code according to a set of rules and heuristics. Like most tools, it is not perfect nor covers every single case, but it is good enough to be helpful.

Signed-off-by: slowy07 <slowy.arfy@gmail.com>